DEV-2021 Transform DC metadata to MHS format
Add CP ID, SP name, md5, source and PID in the transformation via XSLT parameters
as these values are not present in the input XML. Do note that the PID used to be
in the input XML as `dcterms:identifier` but that value is not known at that
moment. The AIP creator actually requests the PID from the PID service.

Also:
 - Remove unused config.
 - Format via black.