DEV-2079 Transform premis information
Certain premis information (= local_id(s)) were incorrectly contained in the
descriptive metadata. This is now moved to the preservation metadata of the
package/ie.

Also, implement a change in the preservation metadata on the representation
level: the value of the `xsi:type` attribute of a `premis:object` node needs
to be prefixed with `premis`, e.g. `<premis:object xsi:type="premis:file">`.