DEV-2173 Map sp_name correctly in case of XDCAM
In all cases, the `sp_name` is added in the mhs metadata with the hardcoded
value `sipin`. In the case of XDCAM, the `sp_name` is supplied in the
sidecar metadata which is also added in the mhs metadata. In short, in that
case, the `sp_name` element is added twice with different values.

The `sp_name` should be calculated:
 - Add the value from the supplied metadata if available;
 - Otherwise use the hardcoded value `sipin`.